+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] 
+### Added
+- **PODAAC-4324**
+  - Make output of this lambda compliant with cumulus 10.1.2 schema
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
 ## [v1.7.0] - 2022-03-15
 ### Added
 - **PODAAC-4308**